requirements.txt
Pillow==10.2.0
pyautogui==0.9.54
datasets
transformers
evaluate
timm
albumentations
accelerate
matplotlib
seaborn
ipywidgets
pycocotools
opencv-python
easyocr
Flask==3.0.2
